Refactored constants; renamed constants in webmon DBUtil now actually clears the PQresult DBUtil now looks at query to determine if it is multi-command. If it is multi-command, it will use PQexec. Otherwise, it will use PQexecparameters. This fixes a bug where garbage was returned when passing 0 parameters to a DBUtil::query call Renamed Drawer to Plotter in webmon Adjusted and renamed frameloader to be able to load frames from the newly-written Prosurd. Adjusted and renamed SessionList, now JobList, to list jobs using the newly-written Prosurd. Removed LegendUpdater, incorporated into plotter In Webmon in Main, reorganized many variables into namespaces, also moved things around between Main and Plotter In Wenmon, switched theme to dark In Webmon, deleted Tester

+++ prosurd/src/Webserver/Resources/Jobs.cpp
@@ -23,9 +23,9 @@ namespace Prosur::Webserver::Resources::Jobs{
 		json returnObject;
 		for(auto& row: jobs){
 			json job;
-			job["time"] = (int64_t) row["time"];
+			job["time"] = row["time"];
 			job["job_file_name"] = row["job_file_name"];
-			job["job_id"] = (int) row["job_id"];
+			job["job_id"] = row["job_id"];
 			returnObject.push_back(job);
 		}
 		responseBody = returnObject;
